requirements.txt
email-validator>=2.2.0
flask-login>=0.6.3
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
flask-wtf>=1.2.2
openai>=1.65.4
wtforms>=3.2.1
faiss-cpu>=1.10.0
numpy>=1.24.0
sqlalchemy>=2.0.38
line-bot-sdk>=3.16.1 